uncertain hamming loss: fixed wrong data type of uncertain component ... resulting in erroneous calculations later on

+++ src/main/java/com/cs_pum/uncertain_mlc/losses/UncertainHammingLoss.java
@@ -100,7 +100,7 @@ public class UncertainHammingLoss implements MultiLabelLossFunction, Measure {
         double[] probabilities = prediction.getConfidences();
         boolean[] bipartition = prediction.getBipartition();
         double symmetricDifference = 0;
-        int u = 0;
+        double u = 0;
 
         for (int i = 0; i < groundTruth.length; i++) {
             if (probabilities[i] < tao || probabilities[i] > (1 - tao)) {
